+++ axum/src/extract/mod.rs
@@ -6,6 +6,7 @@ use rejection::*;
 
 pub mod connect_info;
 pub mod extractor_middleware;
+pub mod path;
 pub mod rejection;
 
 #[cfg(feature = "ws")]
@@ -16,7 +17,6 @@ mod content_length_limit;
 mod extension;
 mod form;
 mod matched_path;
-mod path;
 mod query;
 mod raw_query;
 mod request_parts;