+++ axum/src/util.rs
@@ -1,5 +1,6 @@
 use bytes::Bytes;
 use pin_project_lite::pin_project;
+use std::fmt;
 use std::ops::Deref;
 
 /// A string like type backed by `Bytes` making it cheap to clone.
@@ -27,6 +28,12 @@ impl ByteStr {
         // `ByteStr` can only be constructed from strings which are always valid
         // utf-8 so this wont panic.
         std::str::from_utf8(&self.0).unwrap()
+    }
+}
+
+impl fmt::Display for ByteStr {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_str())
     }
 }